refactor: remove mocks folder after each test

+++ src/services.ts
@@ -74,9 +74,26 @@ export class FileService {
     await fs.promises.mkdir(folder)
   }
 
-  public async removeFolder(folder: string) {
-    // @ts-ignore
-    await fs.promises.rmdir(folder, { recursive: true })
+  public async removeFolder(folder: string | RegExp) {
+    if (typeof folder === 'string') {
+      // @ts-ignore
+      await fs.promises.rmdir(folder, { recursive: true })
+      return
+    }
+
+    const folders = await this.getFoldersNames('.')
+    const filteredFolders = folders.filter(filteredFolder =>
+      folder.test(filteredFolder)
+    )
+    await Promise.all(filteredFolders.map(this.removeFolder))
+  }
+
+  public async getFoldersNames(rootFolder: string): Promise<string[]> {
+    const files = await fs.promises.readdir(rootFolder, { withFileTypes: true })
+
+    return files
+      .filter(dirent => dirent.isDirectory())
+      .map(dirent => dirent.name)
   }
 }
